Splitting GitModel to have a view-independant display of modifications.
Previously, the display of the model's modifications was toggled in the
model with self._show_modifications. This can be a problem when trying
to be view-independant. Therefore, we're splitting the model into an
editable model and a not editable model.

Since a lot of code was common to the both models (especially data()),
we're extending GitModel to create GitEditableModel.