Major refactoring of the way we fetch orig states.
The orig state fetching needed to be refactored because we were reading
a Tree that belonged to the old base branch. Meaning that if you added a
file in a commit A (inserted commit), then modified some other file in a
commit B (base commit), then picked a commit C that changed the first
file, we were asking commit B to give us the orig state. The right thing
to do would have been to ask B', the rewritten commit, but that would
have been quite difficult. Instead, we're resetting the temporary rebase
branch, and fetching directly the content of the files.

Also, renaming self._u_files to a more explicit self._statuses_and_files