Temporary fix of race condition.
Because the population of the model (in git_rebase_process) and the
reset of the q_git_model (in gitbuster.progress_thread) are not
synchronized, we may run into this race condition. This will be fixed
later by moving the progress_thread in the q_editable_git_model and
letting it populate and reset the model.