Using the new git_filter_rebase class. This is an important commit.
Before this commit, we had the choice between using git_filter_process,
a frontend for the git filter-branch process, which was able to rewrite
metadata but couldn't deal with inserted or deleted commits. After that
we coded git_rebase_process, a frontend for the git cherry-pick command
mainly, which was able to add or delete commits from a branch. The
problem was that it didn't take into account merge commits and merge
subtrees (same behaviour as git rebase -i).

This new git_filter_rebase class makes good use of git cherry-pick, git
write-tree, git commit-tree, to preserve the merged subtrees of the
branch.

We may still need to go over some test cases containing conflicting
merge commits, but this is a great step for gfbi_core.