Tracking the children of the commits the same way we track the parents.
This surely adds a time overhead, but we need it to be able to set
parents correctly when dealing with non linear histories. Linear
histories makes it easy to insert a commit, because the commit above is
the children commit and the one under is the parent. With non linear
histories, this can't be applied, we need to be more meticulous.