Reverting f8c9d85d4886b77e05d3d9319e8336390ee05886.
Not ignoring children column was a bad solution. The children column
doesn't represent real modifications of the commit objects. The real
solution was to use deleted commits to know what commits should be
rewritten.

(for information, the problem was that deleting one of the initial
commit's children would mark the first commit has modified and it would
make the git_rebase_filter process fail when trying to find the parents
of the initial commit)